+++ com.dotsui.core/DotsUI.Core.Tests/RectTransformTests.cs
@@ -5,6 +5,7 @@ using NUnit.Framework;
 using Unity.Collections;
 using Unity.Entities;
 using Unity.Entities.Tests;
+using Unity.Mathematics;
 using Unity.Transforms;
 
 namespace DotsUI.Core.Tests
@@ -20,9 +21,9 @@ namespace DotsUI.Core.Tests
             base.Setup();
 
             m_ChildArchetype = m_Manager.CreateArchetype(ComponentType.ReadWrite<RectTransform>(),
-                ComponentType.ReadWrite<Parent>());
+                ComponentType.ReadWrite<Parent>(), typeof(LocalToWorld), typeof(LocalToParent));
             m_RootArchetype = m_Manager.CreateArchetype(ComponentType.ReadWrite<RectTransform>(),
-                ComponentType.ReadWrite<WorldSpaceRect>());
+                ComponentType.ReadWrite<WorldSpaceRect>(), typeof(LocalToWorld));
         }
         /*
          * Test hierarchy:
@@ -41,6 +42,8 @@ namespace DotsUI.Core.Tests
         {
             var root = m_Manager.CreateEntity(m_RootArchetype);
             var childArchetype = m_Manager.CreateArchetype(ComponentType.ReadWrite<RectTransform>(),
+                ComponentType.ReadWrite<LocalToParent>(),
+                ComponentType.ReadWrite<LocalToWorld>(),
                 ComponentType.ReadWrite<WorldSpaceRect>(),
                 ComponentType.ReadWrite<Parent>());
             NativeArray<Entity> children = new NativeArray<Entity>(8, Allocator.Temp, NativeArrayOptions.UninitializedMemory);
@@ -53,7 +56,7 @@ namespace DotsUI.Core.Tests
             m_Manager.SetComponentData(children[5], new Parent() { Value = children[4] });
             m_Manager.SetComponentData(children[6], new Parent() { Value = children[5] });
             m_Manager.SetComponentData(children[7], new Parent() { Value = root });
-            World.GetOrCreateSystem<ParentSystem>().Update();
+            World.GetOrCreateSystem<EndFrameParentSystem>().Update();
             return (root, children);
         }
 
@@ -79,7 +82,7 @@ namespace DotsUI.Core.Tests
 
             var child = m_Manager.CreateEntity(m_ChildArchetype);
             m_Manager.SetComponentData(child, new Parent() { Value = children[7] });
-            World.GetOrCreateSystem<ParentSystem>().Update();
+            World.GetOrCreateSystem<EndFrameParentSystem>().Update();
 
             Assert.AreEqual(child, m_Manager.GetBuffer<Child>(children[7])[0].Value);
         }
@@ -90,7 +93,7 @@ namespace DotsUI.Core.Tests
             var (root, children) = CreateInitialHierarchy();
 
             m_Manager.RemoveComponent<Parent>(children[5]);
-            World.GetOrCreateSystem<ParentSystem>().Update();
+            World.GetOrCreateSystem<EndFrameParentSystem>().Update();
             Assert.Throws<ArgumentException>(() => m_Manager.GetBuffer<Child>(children[4]));
         }
 
@@ -100,7 +103,7 @@ namespace DotsUI.Core.Tests
             var (root, children) = CreateInitialHierarchy();
 
             m_Manager.SetComponentData(children[6], new Parent() { Value = children[4] });
-            World.GetOrCreateSystem<ParentSystem>().Update();
+            World.GetOrCreateSystem<EndFrameParentSystem>().Update();
             Assert.Throws<ArgumentException>(() => m_Manager.GetBuffer<Child>(children[5]));
             Assert.AreEqual(2, m_Manager.GetBuffer<Child>(children[4]).Length);
         }
@@ -114,6 +117,78 @@ namespace DotsUI.Core.Tests
         [Test]
         public void TestPhysicalScaledCanvas()
         {
+        }
+
+        [Test]
+        public void InverseRectTransform()
+        {
+            NativeArray<WorldSpaceRect> parentRects = new NativeArray<WorldSpaceRect>(4, Allocator.Temp);
+            parentRects[0] = new WorldSpaceRect() { Min = new float2(0.0f, 0.0f), Max = new float2(100.0f, 100.0f) };
+            parentRects[1] = new WorldSpaceRect() { Min = new float2(10.0f, 10.0f), Max = new float2(100.0f, 100.0f) };
+            parentRects[2] = new WorldSpaceRect() { Min = new float2(-10.0f, -10.0f), Max = new float2(100.0f, 100.0f) };
+            parentRects[3] = new WorldSpaceRect() { Min = new float2(-100.0f, -100.0f), Max = new float2(-10.0f, -10.0f) };
+            NativeArray<RectTransform> rectTransforms = new NativeArray<RectTransform>(4, Allocator.Temp);
+            rectTransforms[0] = new RectTransform()
+            {
+                AnchorMin = new float2(0.0f, 0.0f),
+                AnchorMax = new float2(1.0f, 1.0f),
+                Pivot = new float2(0.5f, 0.5f),
+                Position = new float2(5.0f, 5.0f),
+                SizeDelta = new float2(20.0f, 20.0f)
+            };
+            rectTransforms[1] = new RectTransform()
+            {
+                AnchorMin = new float2(0.0f, 0.0f),
+                AnchorMax = new float2(0.0f, 1.0f),
+                Pivot = new float2(1.0f, 1.0f),
+                Position = new float2(5.0f, 5.0f),
+                SizeDelta = new float2(20.0f, 20.0f)
+            };
+            rectTransforms[2] = new RectTransform()
+            {
+                AnchorMin = new float2(0.5f, 0.5f),
+                AnchorMax = new float2(0.5f, 0.5f),
+                Pivot = new float2(0.5f, 0.5f),
+                Position = new float2(-5.0f, -5.0f),
+                SizeDelta = new float2(20.0f, 20.0f)
+            };
+            rectTransforms[3] = new RectTransform()
+            {
+                AnchorMin = new float2(1.0f, 1.0f),
+                AnchorMax = new float2(1.0f, 1.0f),
+                Pivot = new float2(0.0f, 0.0f),
+                Position = new float2(-5.0f, -5.0f),
+                SizeDelta = new float2(20.0f, 20.0f)
+            };
+            for (int i = 0; i < parentRects.Length; i++)
+            {
+                for (int j = 0; j < rectTransforms.Length; j++)
+                {
+                    var worldSpace = RectTransformUtils.CalculateWorldSpaceRect(parentRects[i], new float2(1.0f, 1.0f), rectTransforms[j]);
+                    var emptyTransform = new RectTransform()
+                    {
+                        AnchorMin = rectTransforms[j].AnchorMin,
+                        AnchorMax = rectTransforms[j].AnchorMax,
+                        Pivot = rectTransforms[j].Pivot
+                    };
+                    var calculatedTransform = RectTransformUtils.CalculateInverseTransformWithAnchors(worldSpace,
+                        parentRects[i], emptyTransform, new float2(1.0f, 1.0f));
+
+                    Assert.IsTrue(RectTransformEquals(rectTransforms[j], calculatedTransform));
+                }
+            }
+        }
+
+        bool RectTransformEquals(RectTransform item1, RectTransform item2)
+        {
+            return Float2Equals(item1.AnchorMin, item2.AnchorMin) && Float2Equals(item1.AnchorMax, item2.AnchorMax) &&
+                   Float2Equals(item1.Pivot, item2.Pivot) && Float2Equals(item1.Position, item2.Position) &&
+                   Float2Equals(item1.SizeDelta, item2.SizeDelta);
+        }
+
+        bool Float2Equals(float2 item1, float2 item2)
+        {
+            return math.abs(item1.x - item2.x) < 0.001f && math.abs(item1.y - item2.y) < 0.001f;
         }
     }
 }