+++ com.dotsui.core/DotsUI.Core/RectTransformUtils.cs
@@ -65,16 +65,16 @@ namespace DotsUI.Core
         public static WorldSpaceRect CalculateWorldSpaceRect(WorldSpaceRect parentRect, float2 scale,
             RectTransform childTransform)
         {
-            float2 start =
+            var start =
                 math.lerp(parentRect.Min, parentRect.Max,
                     math.lerp(childTransform.AnchorMin, childTransform.AnchorMax, childTransform.Pivot)) +
                 childTransform.Position * scale;
-            float2 anchorDiff = childTransform.AnchorMax - childTransform.AnchorMin;
-            float2 size = (parentRect.Max - parentRect.Min) * anchorDiff +
+            var anchorDiff = childTransform.AnchorMax - childTransform.AnchorMin;
+            var size = (parentRect.Max - parentRect.Min) * anchorDiff +
                           childTransform.SizeDelta * scale;
 
-            float2 min = start - size * childTransform.Pivot;
-            float2 max = start + size * (new float2(1.0f, 1.0f) - childTransform.Pivot);
+            var min = start - size * childTransform.Pivot;
+            var max = start + size * (new float2(1.0f, 1.0f) - childTransform.Pivot);
 
             var childRect = new WorldSpaceRect()
             {
@@ -82,6 +82,15 @@ namespace DotsUI.Core
                 Max = max,
             };
             return childRect;
+        }
+        public static RectTransform CalculateInverseTransformWithAnchors(WorldSpaceRect desiredRect, WorldSpaceRect parent, RectTransform currentTransform, float2 scale)
+        {
+            var anchorDiff = currentTransform.AnchorMax - currentTransform.AnchorMin;
+            var parentSizeRelation = parent.Size * anchorDiff;
+            currentTransform.SizeDelta = (desiredRect.Size - parentSizeRelation) / scale;
+            var start = desiredRect.Min + desiredRect.Size * currentTransform.Pivot;
+            currentTransform.Position = (start - math.lerp(parent.Min, parent.Max, math.lerp(currentTransform.AnchorMin, currentTransform.AnchorMax, currentTransform.Pivot))) / scale;
+            return currentTransform;
         }
     }
 }